Switch for IntEnum to Enum
The reason for the incompatibility of the runner with python 3.11 is an update of the IntEnum Class. 
The str() function now no longer return IntEnum.VALUE but the the integer repressentation of the value. Thats a problem for album_logging.py. Since the function to_loglevel("INFO") now returns 0 and instead of Loglevel.INFO